Include stopwords in II
